pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lodestar"
version = "0.1.0"
description = "Automated trading application with market research, multiple strategies, and a web dashboard"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "BSL-1.0"}
dependencies = [
    "robin_stocks>=3.0.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "ta>=0.11",           # Technical analysis library
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "textblob>=0.18",     # Sentiment analysis
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "jinja2>=3.1",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "rich>=13.0",         # CLI formatting
    "click>=8.1",
    "schedule>=1.2",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
]

[project.scripts]
lodestar = "lodestar.cli.main:cli"

[tool.setuptools.packages.find]
include = ["lodestar*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
